fix(swipe): iOS Safari対応 - ネイティブイベントリスナー + touch-action修正
## 問題
iPhoneでスワイプナビゲーションが動作しない

## 原因
1. React SyntheticEventは passive: false を設定できない
2. iOS SafariではpassiveイベントでpreventDefault()が効かない
3. touch-action: pan-y が水平スワイプをブロック

## 対策
1. useSwipe: ネイティブイベントリスナー（passive: false）に移行
2. FishingRecordDetail/FishingMap: touch-action を pan-x pan-y に変更
3. handlers を空オブジェクトに（後方互換性維持、@deprecated）

## テスト
- TestSwipeComponent を使用した実DOM対応テスト
- PointerEventポリフィル追加（JSDOM対応）
- .ts → .tsx にリネーム（JSX使用のため）

Closes #365
