--- src/hooks/__tests__/useSwipe.test.ts
@@ -1,621 +1,0 @@
-/**
- * useSwipeカスタムフックとswipe-utilsの単体テスト
- *
- * @description
- * スワイプナビゲーション機能のテストスイート
- *
- * @version 1.0.0
- * @since 2025-12-03 Issue #365
- */
-
-import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
-import { renderHook, act, waitFor } from '@testing-library/react';
-import { useSwipe } from '../useSwipe';
-import {
-  haversineDistance,
-  getSwipeDirection,
-  calculateSwipeVelocity,
-  isInEdgeZone,
-  isVerticalDeviationAllowed,
-  calculateSwipeProgress,
-  prefersReducedMotion,
-  DEFAULT_SWIPE_CONFIG,
-} from '../../lib/swipe-utils';
-
-// matchMedia mock
-const mockMatchMedia = (matches: boolean) => {
-  return vi.fn().mockImplementation((query) => ({
-    matches,
-    media: query,
-    onchange: null,
-    addListener: vi.fn(),
-    removeListener: vi.fn(),
-    addEventListener: vi.fn(),
-    removeEventListener: vi.fn(),
-    dispatchEvent: vi.fn(),
-  }));
-};
-
-// navigator.vibrate mock
-const mockVibrate = vi.fn();
-
-describe('swipe-utils', () => {
-  describe('haversineDistance', () => {
-    it('calculates distance between two points correctly', () => {
-      // 東京駅 → 渋谷駅（約3.3km）
-      const tokyo = { latitude: 35.6812, longitude: 139.7671 };
-      const shibuya = { latitude: 35.6586, longitude: 139.7454 };
-
-      const distance = haversineDistance(tokyo, shibuya);
-
-      // 約3300メートル（誤差10%許容）
-      expect(distance).toBeGreaterThan(2900);
-      expect(distance).toBeLessThan(3700);
-    });
-
-    it('returns 0 for same coordinates', () => {
-      const point = { latitude: 35.6812, longitude: 139.7671 };
-      const distance = haversineDistance(point, point);
-      expect(distance).toBe(0);
-    });
-
-    it('calculates short distance correctly', () => {
-      // 約100メートルの距離
-      const point1 = { latitude: 35.6812, longitude: 139.7671 };
-      const point2 = { latitude: 35.6812, longitude: 139.7683 }; // 経度のみ微増
-
-      const distance = haversineDistance(point1, point2);
-      expect(distance).toBeGreaterThan(50);
-      expect(distance).toBeLessThan(200);
-    });
-
-    it('calculates long distance correctly', () => {
-      // 東京 → ニューヨーク（約10,800km）
-      const tokyo = { latitude: 35.6812, longitude: 139.7671 };
-      const newYork = { latitude: 40.7128, longitude: -74.006 };
-
-      const distance = haversineDistance(tokyo, newYork);
-
-      // 約10,800km（誤差5%許容）
-      expect(distance).toBeGreaterThan(10_000_000);
-      expect(distance).toBeLessThan(12_000_000);
-    });
-  });
-
-  describe('getSwipeDirection', () => {
-    it('returns "right" for positive deltaX above threshold', () => {
-      expect(getSwipeDirection(100, 80)).toBe('right');
-    });
-
-    it('returns "left" for negative deltaX above threshold', () => {
-      expect(getSwipeDirection(-100, 80)).toBe('left');
-    });
-
-    it('returns "none" for deltaX below threshold', () => {
-      expect(getSwipeDirection(50, 80)).toBe('none');
-      expect(getSwipeDirection(-50, 80)).toBe('none');
-    });
-
-    it('returns "none" for deltaX equal to threshold', () => {
-      // 閾値未満のため none
-      expect(getSwipeDirection(79, 80)).toBe('none');
-    });
-  });
-
-  describe('calculateSwipeVelocity', () => {
-    it('calculates velocity correctly', () => {
-      expect(calculateSwipeVelocity(100, 200)).toBe(0.5); // 100px / 200ms = 0.5 px/ms
-    });
-
-    it('returns 0 for zero duration', () => {
-      expect(calculateSwipeVelocity(100, 0)).toBe(0);
-    });
-
-    it('handles negative deltaX', () => {
-      expect(calculateSwipeVelocity(-100, 200)).toBe(0.5);
-    });
-  });
-
-  describe('isInEdgeZone', () => {
-    it('returns true for clientX within edge zone', () => {
-      expect(isInEdgeZone(10, 16)).toBe(true);
-      expect(isInEdgeZone(16, 16)).toBe(true);
-    });
-
-    it('returns false for clientX outside edge zone', () => {
-      expect(isInEdgeZone(17, 16)).toBe(false);
-      expect(isInEdgeZone(100, 16)).toBe(false);
-    });
-  });
-
-  describe('isVerticalDeviationAllowed', () => {
-    it('returns true for deviation within limit', () => {
-      expect(isVerticalDeviationAllowed(30, 50)).toBe(true);
-      expect(isVerticalDeviationAllowed(-30, 50)).toBe(true);
-      expect(isVerticalDeviationAllowed(50, 50)).toBe(true);
-    });
-
-    it('returns false for deviation exceeding limit', () => {
-      expect(isVerticalDeviationAllowed(51, 50)).toBe(false);
-      expect(isVerticalDeviationAllowed(-51, 50)).toBe(false);
-    });
-  });
-
-  describe('calculateSwipeProgress', () => {
-    it('returns progress between 0 and 1', () => {
-      expect(calculateSwipeProgress(0, 80)).toBe(0);
-      expect(calculateSwipeProgress(40, 80)).toBe(0.5);
-      expect(calculateSwipeProgress(80, 80)).toBe(1);
-    });
-
-    it('clamps progress at 1 for values exceeding threshold', () => {
-      expect(calculateSwipeProgress(160, 80)).toBe(1);
-    });
-
-    it('handles negative deltaX', () => {
-      expect(calculateSwipeProgress(-40, 80)).toBe(0.5);
-    });
-  });
-
-  describe('prefersReducedMotion', () => {
-    beforeEach(() => {
-      window.matchMedia = mockMatchMedia(false);
-    });
-
-    it('returns false when user prefers motion', () => {
-      expect(prefersReducedMotion()).toBe(false);
-    });
-
-    it('returns true when user prefers reduced motion', () => {
-      window.matchMedia = mockMatchMedia(true);
-      expect(prefersReducedMotion()).toBe(true);
-    });
-  });
-
-  describe('DEFAULT_SWIPE_CONFIG', () => {
-    it('has correct default values', () => {
-      expect(DEFAULT_SWIPE_CONFIG.DETAIL_THRESHOLD).toBe(80);
-      expect(DEFAULT_SWIPE_CONFIG.POPUP_THRESHOLD).toBe(60);
-      expect(DEFAULT_SWIPE_CONFIG.DETAIL_VELOCITY_THRESHOLD).toBe(0.3);
-      expect(DEFAULT_SWIPE_CONFIG.POPUP_VELOCITY_THRESHOLD).toBe(0.25);
-      expect(DEFAULT_SWIPE_CONFIG.EDGE_ZONE).toBe(16);
-      expect(DEFAULT_SWIPE_CONFIG.NEARBY_DISTANCE_THRESHOLD).toBe(5000);
-    });
-  });
-});
-
-describe('useSwipe', () => {
-  beforeEach(async () => {
-    window.matchMedia = mockMatchMedia(false);
-    Object.defineProperty(navigator, 'vibrate', {
-      value: mockVibrate,
-      writable: true,
-    });
-
-    // CI環境でのJSDOM初期化待機
-    if (process.env.CI) {
-      await waitFor(
-        () => {
-          if (!document.body || document.body.children.length === 0) {
-            throw new Error('JSDOM not ready');
-          }
-        },
-        { timeout: 5000, interval: 100 }
-      );
-    } else {
-      await new Promise((resolve) => setTimeout(resolve, 0));
-    }
-  });
-
-  afterEach(() => {
-    vi.restoreAllMocks();
-    vi.useRealTimers();
-    mockVibrate.mockClear();
-    if (!process.env.CI) {
-      document.body.innerHTML = '';
-    }
-  });
-
-  describe('Hook initialization', () => {
-    it('returns ref, state, reset, and handlers', () => {
-      const { result } = renderHook(() => useSwipe());
-
-      expect(result.current.ref).toBeDefined();
-      expect(result.current.state).toBeDefined();
-      expect(result.current.reset).toBeDefined();
-      expect(result.current.handlers).toBeDefined();
-      expect(typeof result.current.handlers.onPointerDown).toBe('function');
-      expect(typeof result.current.handlers.onPointerMove).toBe('function');
-      expect(typeof result.current.handlers.onPointerUp).toBe('function');
-      expect(typeof result.current.handlers.onPointerCancel).toBe('function');
-    });
-
-    it('initializes with correct default state', () => {
-      const { result } = renderHook(() => useSwipe());
-
-      expect(result.current.state.isSwiping).toBe(false);
-      expect(result.current.state.progress).toBe(0);
-      expect(result.current.state.direction).toBe('none');
-      expect(result.current.state.offsetX).toBe(0);
-      expect(result.current.state.thresholdReached).toBe(false);
-    });
-  });
-
-  describe('Swipe callbacks', () => {
-    it('calls onSwipeStart when swipe begins', () => {
-      const onSwipeStart = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({}, { onSwipeStart })
-      );
-
-      // Create test element
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const mockEvent = createMockPointerEvent(100, 100);
-
-      act(() => {
-        result.current.handlers.onPointerDown(mockEvent);
-      });
-
-      expect(onSwipeStart).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls onSwipeProgress during swipe', () => {
-      const onSwipeProgress = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ threshold: 80 }, { onSwipeProgress })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(100, 100);
-      const moveEvent = createMockPointerEvent(140, 100); // 40px right
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerMove(moveEvent);
-      });
-
-      expect(onSwipeProgress).toHaveBeenCalled();
-      const [progress, direction] = onSwipeProgress.mock.calls[0];
-      expect(progress).toBe(0.5); // 40 / 80 = 0.5
-      expect(direction).toBe('right');
-    });
-
-    it('calls onSwipeLeft when swiping left past threshold', async () => {
-      vi.useFakeTimers();
-      const onSwipeLeft = vi.fn();
-      const onSwipeEnd = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ threshold: 80, animationDuration: 300 }, { onSwipeLeft, onSwipeEnd })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(200, 100);
-      const upEvent = createMockPointerEvent(100, 100); // 100px left
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerUp(upEvent);
-      });
-
-      // アニメーション完了を待つ
-      await act(async () => {
-        vi.advanceTimersByTime(400);
-      });
-
-      expect(onSwipeLeft).toHaveBeenCalledTimes(1);
-    });
-
-    it('calls onSwipeRight when swiping right past threshold', async () => {
-      vi.useFakeTimers();
-      const onSwipeRight = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ threshold: 80, animationDuration: 300 }, { onSwipeRight })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(100, 100);
-      const upEvent = createMockPointerEvent(200, 100); // 100px right
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerUp(upEvent);
-      });
-
-      await act(async () => {
-        vi.advanceTimersByTime(400);
-      });
-
-      expect(onSwipeRight).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('Edge zone handling', () => {
-    it('ignores swipe starting in edge zone', () => {
-      const onSwipeStart = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ edgeZone: 16 }, { onSwipeStart })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      // Start in edge zone (x=10, edge=16)
-      const mockEvent = createMockPointerEvent(10, 100);
-
-      act(() => {
-        result.current.handlers.onPointerDown(mockEvent);
-      });
-
-      expect(onSwipeStart).not.toHaveBeenCalled();
-      expect(result.current.state.isSwiping).toBe(false);
-    });
-
-    it('accepts swipe starting outside edge zone', () => {
-      const onSwipeStart = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ edgeZone: 16 }, { onSwipeStart })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      // Start outside edge zone (x=20, edge=16)
-      const mockEvent = createMockPointerEvent(20, 100);
-
-      act(() => {
-        result.current.handlers.onPointerDown(mockEvent);
-      });
-
-      expect(onSwipeStart).toHaveBeenCalledTimes(1);
-      expect(result.current.state.isSwiping).toBe(true);
-    });
-  });
-
-  describe('Vertical deviation handling', () => {
-    it('cancels swipe when vertical deviation exceeds limit', () => {
-      const onSwipeProgress = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ maxVerticalDeviation: 50 }, { onSwipeProgress })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(100, 100);
-      const moveEvent = createMockPointerEvent(150, 200); // 100px vertical deviation
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerMove(moveEvent);
-      });
-
-      // Swipe should be cancelled
-      expect(result.current.state.isSwiping).toBe(false);
-    });
-  });
-
-  describe('Disabled directions', () => {
-    it('applies damping when left swipe is disabled', () => {
-      const { result } = renderHook(() =>
-        useSwipe({ threshold: 80, disableLeft: true, dampingFactor: 0.3 })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(200, 100);
-      const moveEvent = createMockPointerEvent(100, 100); // 100px left
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerMove(moveEvent);
-      });
-
-      // Damped offset should be applied
-      expect(result.current.state.offsetX).toBe(-100 * 0.3); // -30px
-      expect(result.current.state.progress).toBe(0);
-    });
-
-    it('calls onEdgeReached when swiping in disabled direction', async () => {
-      vi.useFakeTimers();
-      const onEdgeReached = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe({ threshold: 80, disableLeft: true }, { onEdgeReached })
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(200, 100);
-      const upEvent = createMockPointerEvent(100, 100); // 100px left (disabled)
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerUp(upEvent);
-      });
-
-      await act(async () => {
-        vi.advanceTimersByTime(500);
-      });
-
-      expect(onEdgeReached).toHaveBeenCalledWith('left');
-    });
-  });
-
-  describe('Reset functionality', () => {
-    it('resets state to initial values', () => {
-      const { result } = renderHook(() => useSwipe());
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(100, 100);
-      const moveEvent = createMockPointerEvent(150, 100);
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        result.current.handlers.onPointerMove(moveEvent);
-      });
-
-      // State should be updated
-      expect(result.current.state.isSwiping).toBe(true);
-
-      // Reset
-      act(() => {
-        result.current.reset();
-      });
-
-      expect(result.current.state.isSwiping).toBe(false);
-      expect(result.current.state.progress).toBe(0);
-      expect(result.current.state.direction).toBe('none');
-      expect(result.current.state.offsetX).toBe(0);
-    });
-  });
-
-  describe('Pointer cancel handling', () => {
-    it('resets state on pointer cancel', () => {
-      const { result } = renderHook(() => useSwipe());
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      const downEvent = createMockPointerEvent(100, 100);
-      const cancelEvent = createMockPointerEvent(150, 100);
-
-      act(() => {
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      expect(result.current.state.isSwiping).toBe(true);
-
-      act(() => {
-        result.current.handlers.onPointerCancel(cancelEvent);
-      });
-
-      // アニメーション後にリセットされる
-      // Note: 実際にはアニメーション完了後にリセットされるが、テストでは即時確認
-    });
-  });
-
-  describe('Velocity-based swipe detection', () => {
-    it('triggers swipe on high velocity even below threshold', async () => {
-      vi.useFakeTimers();
-      const onSwipeRight = vi.fn();
-      const { result } = renderHook(() =>
-        useSwipe(
-          { threshold: 80, velocityThreshold: 0.3 },
-          { onSwipeRight }
-        )
-      );
-
-      const element = document.createElement('div');
-      Object.defineProperty(result.current.ref, 'current', {
-        value: element,
-        writable: true,
-      });
-
-      // 50px in 100ms = 0.5 px/ms (above velocity threshold)
-      const downEvent = createMockPointerEvent(100, 100, 0);
-      const upEvent = createMockPointerEvent(150, 100, 100);
-
-      act(() => {
-        vi.setSystemTime(0);
-        result.current.handlers.onPointerDown(downEvent);
-      });
-
-      act(() => {
-        vi.setSystemTime(100);
-        result.current.handlers.onPointerUp(upEvent);
-      });
-
-      await act(async () => {
-        vi.advanceTimersByTime(400);
-      });
-
-      expect(onSwipeRight).toHaveBeenCalledTimes(1);
-    });
-  });
-});
-
-/**
- * モックポインターイベントを作成
- */
-function createMockPointerEvent(
-  clientX: number,
-  clientY: number,
-  _timestamp?: number
-): React.PointerEvent<HTMLElement> {
-  const mockTarget = document.createElement('div');
-
-  return {
-    clientX,
-    clientY,
-    pointerId: 1,
-    target: mockTarget,
-    currentTarget: mockTarget,
-    preventDefault: vi.fn(),
-    stopPropagation: vi.fn(),
-  } as unknown as React.PointerEvent<HTMLElement>;
-}
